Rename `form2` directory
Before we pull `Form2` out to a public repo, we want to give it a proper
name. The original paper by Cooper et al., which served as inspiration
for this library, called their abstraction `Formlet`, and the name was
suggested by Arthur, who authored the majority of this library, so we go
with that name.

We start by renaming the `form2` directory.